Made a small change to the collision detection between the ball and the paddle
I changed the if statement on line 187 so that if the bottom of the
ball hit the paddle, the ball would change direction.
I could do the same type of thing to the walls, but I wanted to make
sure that I’m not fixing something that ain’t broken first.